Added basic Pane Component; reorganized the code
The project structure will be as follows: per vite index.html is the
entry point, and technically we can import any scripts we want(using
relative imports) within our project. Within Src I'd like for us to
divide our code into modules. We have a `src/Components/` which will be
similar to view if you think about MVP, and from here we can put code
that "constructs view elements" First Component module is Panes, and we
have minimal viable class for panes created, that can generate into view
either a string or HTML content. Following this will be the PaneManager,
which will store an array of panes and manage Their layout within the
home project. Ensure that the number of panes configured properly
utilize the total screen space available without exceeding it and allow
for reordering and the like.

Another module we've added here is the Statemanager, and code like this,
project level services will go within the `src/services/` directory.
Here we have my EventBus and StateManager. StateManager is built on top
of the event bus and each are managed as independent modules and can be
used throughout the project. State declarations are expected to be done
within, for instance we have here, TestComonet (misspelled), if it's
mainly for a single component or group of components within the
component domain. There may be instances of global state (like
application options for instance), that likely warrant a global state
declaration, so for this we'll likely create `src/States`. Alternatively
if we imagine there being minimum globally scoped state (now mind you
the component level state is still globally accessible, but the project
structure is largely to express who the state mostly belongs to. Other
components, through this implementation can read and even write to this
state, so developer code of conduct may need to be established.

This is just a first pass at project structure. Some things we need next
in basic priority order:
1. Create PanelManager and integrate with Panes; PanelManager will have
a field panes: Panes[], and another field may be useful of Max number of
panes that should be settings driven.
2. Navigation Module.
3. Options and Settings Module
4. Game system manifest.json scoping and implementation for importing GM
screen Data.
5. Markdown to HTML or some way to easily import new panels
6. Create Panels within the application itself with either a markdown
text editor or rich text editor that is able to create new panels right
from the program.
7. Spell database
8. Character sheet manager
9. GM Note taking

Some of this may should be exported into it's own TODO.md, but this will
do for now!